--- src/main/java/umc/lightup/light_talk/domain/Comment.java
@@ -1,33 +1,0 @@
-package umc.lightup.light_talk.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-import umc.lightup.common.BaseEntity;
-import umc.lightup.member.domain.Member;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String content;
-
-    @Builder.Default
-    private Long likes = 0L;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member commentMember;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-}
