+++ src/main/java/umc/lightup/lighttalk/converter/CommentConverter.java
@@ -1,7 +1,7 @@
-package umc.lightup.light_talk.converter;
+package umc.lightup.lighttalk.converter;
 
-import umc.lightup.light_talk.domain.Comment;
-import umc.lightup.light_talk.dto.CommentResponseDTO;
+import umc.lightup.lighttalk.domain.Comment;
+import umc.lightup.lighttalk.dto.CommentResponseDTO;
 
 public class CommentConverter {
