+++ src/main/java/umc/lightup/lighttalk/domain/Comment.java
@@ -1,7 +1,8 @@
-package umc.lightup.light_talk.domain;
+package umc.lightup.lighttalk.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
+import umc.lightup.common.BaseEntity;
 import umc.lightup.member.domain.Member;
 
 @Entity
@@ -9,17 +10,24 @@ import umc.lightup.member.domain.Member;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class CommentLike {
+public class Comment extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id")
-    private Comment comment;
+    @Column(nullable = false)
+    private String content;
+
+    @Builder.Default
+    private Long likes = 0L;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
-    private Member member;
+    private Member commentMember;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
+
 }
