+++ src/main/java/umc/lightup/lighttalk/domain/CommentLike.java
@@ -1,8 +1,7 @@
-package umc.lightup.light_talk.domain;
+package umc.lightup.lighttalk.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import umc.lightup.common.BaseEntity;
 import umc.lightup.member.domain.Member;
 
 @Entity
@@ -10,15 +9,15 @@ import umc.lightup.member.domain.Member;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostLike extends BaseEntity {
+public class CommentLike {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
+    @JoinColumn(name = "comment_id")
+    private Comment comment;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
