+++ src/main/java/umc/lightup/lighttalk/domain/PostImage.java
@@ -1,0 +1,24 @@
+package umc.lightup.lighttalk.domain;
+
+import jakarta.persistence.*;
+import lombok.*;
+import umc.lightup.common.BaseEntity;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+public class PostImage extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "post_id", nullable = false)
+    private Post post;
+
+    @Column(length = 255, nullable = false)
+    private String imageUrl;
+}
