+++ src/main/java/umc/lightup/lighttalk/repository/PostRepository.java
@@ -1,24 +1,26 @@
-package umc.lightup.light_talk.repository;
+package umc.lightup.lighttalk.repository;
 
+import lombok.NonNull;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import umc.lightup.light_talk.domain.Post;
+import umc.lightup.lighttalk.domain.Post;
 
 import java.util.Optional;
 
 @Repository
 public interface PostRepository extends JpaRepository<Post, Long> {
-    @Query("select distinct p from Post p left join fetch p.postComments pc left join fetch pc.commentMember where p.id = :postId")
-    Optional<Post> findByIdWithCommentsAndCommentMembers(@Param("postId") Long postId);
-    @Query("select distinct p from Post p " +
-            "left join fetch p.postMember pm " +
-            "left join fetch p.postImages pi")
-    Page<Post> findAllWithDetails(Pageable pageable);
+
+    @EntityGraph(attributePaths = {"postComments", "postComments.commentMember"})
+    Optional<Post> findPostWithCommentsById(@NonNull Long postId);
+
+    @Query("select p from Post p left join fetch p.postMember pm ")
+    Page<Post> findAllWithMember(Pageable pageable);
 
     @Modifying
     @Query("update Post p set p.likes = p.likes + 1 where p.id = :postId")
